// api/src/dao/balance_sheet.dao.ts
export interface BalanceSheetDAO {
  year: number;
  current_assets: number;
  cash: number;
  receivables: number;
  inventory: number;
  non_current_assets: number;
  loan: number;
  equipment: number;
  goodwill: number;
  long_term_investment: number;
  total_assets: number;
  long_term_debts: number;
}